ANDROID: cf-x86_defconfig: Disable PCH7/PCH9 drivers
These drivers break the PCI initialization on crosvm for unknown
reasons. This is most likely due to an incompatibility between the
pch@1,0 section in arch/x86/dts/qemu-x86_i440fx.dts and the way crosvm
emulates the PCH, but it will need to be investigated later. As these
drivers are not required for our use cases, we can safely disable them.

Bug: 153027511
Signed-off-by: Ram Muthiah <rammuthiah@google.com>
Change-Id: I1149a1092607ac34e6c1872330842916a75f6b3a