ANDROID: Add CONFIG_PCI_ENUM_ONLY to Kconfig
This allows boards to opt into PCI_ENUM_ONLY without having to directly
specify it in board headers, which is much cleaner.

On crosvm the PCI bus is already initialized, and attempting to re-write
the PCI config space with new IDs fails because it is unsupported, which
confuses U-Boot. Specifying PCI_ENUM_ONLY works around this problem.

With this change, the virtio_blk driver finally starts working.

Bug: 153027511
Signed-off-by: Ram Muthiah <rammuthiah@google.com>
Change-Id: I3df31e2b56dc7ae6d6673d5b57223552fd936806