ANDROID: Add CONFIG_PCI_ENUM_ONLY to Kconfig
This allows boards to opt into PCI_ENUM_ONLY without having to directly
specify it in board headers, which is much cleaner.

On crosvm the PCI bus is already initialized, and attempting to re-write
the PCI config space with new IDs fails because it is unsupported, which
confuses U-Boot. Specifying PCI_ENUM_ONLY works around this problem.

With this change, the virtio_blk driver finally starts working.

Bug: 153027511
Bug: 169336575
Signed-off-by: Alistair Delva <adelva@google.com>
Signed-off-by: Ram Muthiah <rammuthiah@google.com>
Change-Id: I3a3501ceafdce98be39d94825c901ac12f710384