ANDROID: Work around toolchain issue with CONFIG_STATIC_RELA
If U-Boot is built without CONFIG_POSITION_INDEPENDENT, it will run a
tool on the final output to adjust some RELA relocations which doesn't
work with clang/llvm for undiagnosed reasons. These fixes don't seem
to be needed to run U-Boot inside QEMU, so expose the KConfig symbol
so we can disable it in another config.

Bug: 170433295
Change-Id: I0f44119f5b9d8bd17d4efd22af5bd6d719ad718c