Switch Android ARM64 boot flow from bootm to booti
The bootm command is designed to work directly on Android boot images,
but our new android_bootloader flow unpacks the boot image(s) and puts
repacked contents in RAM, so it can be booted as a "raw" image. That
means bootm gets confused and does not work.

The "booti" command supports booting raw ARM64 kernel images; it works
with compressed and uncompressed images, and EFI stubbed images, which
makes it a good choice for us.

Change-Id: I25350675b1917fbced1ef7835b1f9e466c23538f