Remove ANDROID_BOOT_IMAGE handling for ramdisks
By the time booti is called, the android_bootloader code has already
unpacked and relocated the kernel, initrd and cmdline, so we don't need
booti to do this again. This means we can just use the RAW initrd
handling. Remove the additional legacy processing for Android boot
images from the booti path.

Change-Id: Ia26d68484efcb184b149eaabdd6cc4332237fc2e