Align Android ramdisk address to 2MB
The Linux kernel expects the ramdisk to be physical memory page size
aligned, which doesn't match the weird "page size" embedded in the boot
image by default (2048) and which I think refers to flash block size.

So, stop using the boot header page size and hard-code the alignment
to 2MB, which eliminates a WARN_ON on x86 and ARM.

This also handily solves a problem with the booti command on ARM64
which relocates the kernel image by 512K (seemingly for no reason)
which used to overwrite the ramdisk address, but now there is plenty
of space to do this relocation.

Change-Id: I9550ddb92f03ecf3d2f3af2737b32abf3bbcbd50