ANDROID: ARM64 crosvm fdt moved to a fixed location
FDT was located at a RAM size dependent location prior to this change
and a crosvm change - crrev.com/c/2641126

The fdt location no longer needs to be fixed to magic numbers. These
magic numbers are removed and the beginning of uboot is moved to start
immediately after the new fixed fdt location.

Bug: 177926450
Test: local boot of cf arm64 guest on crosmv rocki host

Change-Id: Ibd85c020dd1df958f25aa33c0091c314652ae3cd