ANDROID: Add rk3399 bl31 blob and build script
The trusted-firmware-a project currently does not build with modern
clang, but also cannot be built with Android's old GCC 4.9 toolchain, so
we must build it with host toolchains. These toolchains are not present
on the Android builders so, for now, we must use a binary.

The long term solution here is to make trusted-firmware-a build with
clang and add proper build-from-source integration.

This change is a hack and should not be upstreamed; bl31.elf is a common
name used by other boards/targets, it just happens right now none of
ours need it, so this is safe.

Signed-off-by: Alistair Delva <adelva@google.com>
Change-Id: I9f211e1703e3b5b4c66e3dba4bba5c62715c92b2