ANDROID: Move rockpi4 build to build.sh
Support building U-Boot with build/build.sh, which uses our hermetic
clang toolchain, so the binaries can be built on the Android builders.

This disables STATIC_RELA which is broken with clang at the moment,
similarly to the QEMU targets.

This change also removes any changes we made vs upstream to the
rock-pi-4-rk3399_defconfig and moves them into rockpi4.fragment. This is
merged at build time instead.

Signed-off-by: Alistair Delva <adelva@google.com>
Change-Id: I17e52667e78796aac0557c6a368c0a2b7cabf50f