Add CONFIG_AVB_PUBKEY_FILE to specify the root avb public key
The new config CONFIG_AVB_PUBKEY_FILE is used to specify the path to the
public key file that u-boot uses as the root public key. Previously,
the root public key was hard coded in avb_verify.c and therefore was
non-trivial to replace it with a custom key.

The public key that was hard-coded is extracted as default.avbpubkey
and is used as the default.

Bug: 181924266
Test: DIST_DIR=out/dist BUILD_CONFIG=u-boot/build.config.crosvm.aarch64
build/build.sh -j80

Change-Id: I7d0af97f158b87278a3c635df41e638c918cc96a