AVB BUF in bss section
When the data to be read or written is not on the sector boundaries,
avb_verify.c has used a temporary buffer whose location and size are
specified as CONFIG_AVB_BUF_ADDR and CONFIG_AVB_BUF_SIZE, respective.

This however is cumbersome as, in many cases, the correct buffer
location doesn't really matter. With this change, when
CONFIG_AVB_BUF_ADDR is not set or set to 0, the temporary buffer is
located at bss.

This change also moves some inline functions regarding the temporary
buffer from the header (avb_verify.h) to the source file as they are not
used outside of the source file.

Bug: 181924266
Test: DIST_DIR=out/dist BUILD_CONFIG=u-boot/build.config.crosvm.aarch64
build/build.sh -j80

Change-Id: I8b13d6d022f10a76e328a7b12e7a99d6229725f5