ANDROID: lib: Integrate open-dice code into U-Boot
Integrate the Open Profile for DICE into U-Boot build system, allowing
future AOSP U-Boot code to use that library; only track object files
that are expected to be needed and "include-patch" those as U-Boot does
not provide a proper libc header interface.

Commit soft links to the location of the library, as defined in the AOSP
U-Boot manifest:

    https://android.googlesource.com/kernel/manifest/+/refs/heads/u-boot-mainline

Note: Only the BoringSSL backend for open-dice is supported.

Bug: 198564191
Test: Build with CONFIG_LIB_OPENDICE=y
Change-Id: Ic9da391a172762217c6a24cdbd56ea330dfbbb0c