ANDROID: pvmfw: Fix SYS_{SDRAM,TEXT}_BASE
Update the board config to reflect the address at which crosvm currently
loads the image (SYS_TEXT_BASE) and the RAM base to be exposed to the
kernel (SYS_SDRAM_BASE).

Include the 0x7000_0000-0x7FFF_FFFF range in the MMU mapping as that
range is reserved for FW purpose and should not result in a fault when
modified by pvmfw.

Ensure that build/build.sh doesn't modify SYS_TEXT_BASE.

Bug: 210092587
Test: -
Change-Id: I2b4b848012b460760d7ab52aeceb66f832aecda0