android: read the exact amount of data when loading partitions
Previously, when a piece of data (e.g. boot header, bootconfig, ram
disk, etc.) is loaded into memory, it might have loaded more data than
is requested into memory, corrupting the area next to it. This could
happen when the size of data to be read is not a multiple of block size
(which usually us the case) and the data hasn't yet loaded into memory
during the AVB verification. In such cases, the data is read directly
from the storage using the block I/O APIs, where data is moved in block
unit.

This hasn't caused a problem only because the pieces of data are loaded
in the order of their load addresses. (e.g. vendor_ramdisk, then,
boot_ramdisk, and then bootconfig). However, in a follow-up change, they
won't be loaded in the order.

Fixing the issue by saving the data to be overwritten to a temporary
buffer, and then restoring it back after the block I/O is done. In
addition, unnecessary page alignments on the size of kernel,
vendor_ramdisk and boot_ramdisk; we respect their true sizes that are
written in the boot/vendor_boot headers.

Bug: 211934343
Test: run cf with the modified u-boot
Change-Id: I5742b1ebd265a47741855a259fd6f8b8d0dcb788