Preload boot, vendor_boot, and init_boot partitions to the load_address
Previously, the three partitions are first loaded into malloc'ed buffers
where they are verified, and then copied to their load addresses. This
was a huge waste of memory since the two copies of the partitions were
duplicated in the heap and the load address.

This change fixes the problem by preloading the partitions directly onto
the load addresses. It makes use of the get_preloaded_partition hook
that libavb provides.

Bug: 211934343
Test: run cuttlefish with the modified u-boot

Change-Id: Id7a3ff26742a69d5bfb4f6fbef708a27dc1619e8