ANDROID: Generic Boot Certificate Chain functions
Boot Certification Chain (BCC) is a measured boot protocol built around
the Device Identifier Composition Engine (DICE) specification.
A device that supports BCC/DICE should ensure that each boot stage
performs a round of DICE derivation, generating a new entry in the BCC,
and pass that on to the next stage, so-called BCC handover.

Implement BCC handover using the open-dice library. The caller is
expected to provide the input BCC handover and measurements of all
inputs. As a result, it writes the output BCC handover to a given
buffer. It is up to the caller to wipe the handover memory before
booting further.

Signed-off-by: David Brazdil <dbrazdil@google.com>
Change-Id: Iae631c2e7523102a4f3f813165e344cd1331a62f