ANDROID: boot_android: Integrate Boot Certificate Chain
Extend Android boot flow with the optional BCC handover for measured
boot. Following each round of AVB verification, the boot flow passes AVB
verification data to the BCC implementation, triggering a DICE
derivation and generation of a new BCC handover.

So as to avoid the need for understanding how the BCC handover data is
stored, the routine simply overwrites the incoming handover data.

Signed-off-by: David Brazdil <dbrazdil@google.com>
Change-Id: I4613e5232a4cd808f02f5efce1f872a24e324a0a