ANDROID: Make the semantics of boot_android more useful
Previously, boot_android or verified_boot_android required that a 'misc'
partition was specified to read the boot mode and slot. In some use
cases, such as Microdroid, there was no misc partition, and this was
being incorrectly specified as 'dev', which would fall back to reading
the first partition on the device ('boot_a') as a BCB, which failed, and
fell back to NORMAL boot mode. Since the BCB could not be read, the slot
had to be specified explicitly for boot to succeeed.

This new code removes the ability to specify a partition by numeric
offset, but we had no known users of that path anyway. Now, we can avoid
the complexity of blk_get_device_part_str(), which allowed the invalid
assumptions above to be made without returning a detectable error.

If the user specifies a "dev#part_str", it will be interpreted as fully
compliant Android boot flow, with a specifiable boot mode, and
Android-controled boot slot. If the user only specifies a "dev", it will
assume a boot flow that does not depend on a BCB, and the boot mode will
default to NORMAL and the slot to "_a", unless overridden on the command
line or with the "android_boot_slot" environment variable. This enables
simplified A/B boot flows without requiring a misc partition.

Finally, if a user specifies the whole "dev#part_str" but the "part_str"
partition is missing, it will fall back to the simplified flow and make
some default assumptions. This is useful for unifying our bootloader
environments and builds, once we start building the environment in.

Bug: 216304872
Change-Id: I105f9800a388015e6cac5ad18adf731479765448