ANDROID: Use a built-in environment for all virtual devices
Recent changes to the bootloader have made booting Android much simpler,
with fewer workarounds required. Avoid loading the bootloader
environment from disk, which provides a control vector for an attacker.

This makes the cuttlefish and microdroid bootloaders like kiwi's.

On microdroid, the device "0#misc" will be used, even though this
platform does not have a misc partition. The bootloader will
automatically detect this and fall back to "0", which means the Android
boot flow will assume NORMAL boot is wanted, and the slot suffix should
be "_a".

Bug: 216304872
Change-Id: Ib8dc81000ae42e56b321041c1119630cf698723b