arm64: Add hook to handle synchronous exceptions
In cases where we would want u-boot to handle synchronous exceptions
(either injected by a more provileged exception level, or locally
generated), provide a hook that will be called when such exception
occurs, allowing it to be handled instead of always being fatal.

The default hook is to do nothing and let the exception be fatal.

Signed-off-by: Marc Zyngier <maz@kernel.org>
Change-Id: I6c6f0f4562736be21ccc5bf40a5a27d1ad4a26aa