ANDROID: Set up fdtaddr in board_late_init()
The fdtaddr is the address where an FDT blob is preloaded, and on crosvm
and QEMU the SDRAM base differs, so we can't hard-code it. Putting this
in the arch-neutral default_env.txt doesn't make sense either. Set it up
in the board_late_init() so it happens after the environment has been
initialized. Although it should never need to be overridden, this can
still be done by the bootcmd.

Bug: 216304872
Change-Id: Ieda5b15708c546e36ebb9f860c41471e124e1120