ANDROID: Implement cbootargs for ARM
Implement the ARM version of Ia2d32d3c575c3d88ccb42374e3bb43e4aa1ecee1.

Set up cbootargs in the environment to contain the value of the FDT
/chosen bootargs DT node, if it exists, so it can be used to build a new
bootargs for the kernel boot. U-Boot normally overwrites this node when
it updates the FDT with the bootargs environment variable, unless the
variable is unset. crosvm sets up the chosen node with earlycon, ramoops
and other valuable settings that are otherwise lost.

Bug: 216304872
Change-Id: Ia625462f98797ed7bdd47186555c8b1a37fe3ec0