ANDROID: Remove version from BCC
The version should be of the thing being loaded but is consistently
being set to the version of the thing doing the loading. There's not a
clear way to get the version of the thing being loaded so avoid setting
the wrong version until we find a way to do it right.

Signed-off-by: Andrew Scull <ascull@google.com>
Change-Id: Ie5268aacadea64f75e70f42d099a737f4ae2d8d2