ANDROID: Add helper for VM instance BCC handover
The process to commit boot measurements to the BCC and integrate it with
the VM instance image is similar in each of the cases we have. Add a
helper to avoid duplication of that logic.

A future change will see the compounded hashes be replaced with
structured BCC entries that have beening to and external observer.

Signed-off-by: Andrew Scull <ascull@google.com>
Change-Id: Iaebf4026d8ed622277c4c815c2f62e3f4438ecd7