ANDROID: Refine BCC input choices for AVB partitions
There are two cases to consider for current applications: a verification
of a set of partitions described by a vbmeta partition with no chaining
and loading a single partition that is chained from a vbmeta partition.

Use the complete vbmeta digest when all partitions are being loaded and
the digest of the vbmeta of the single chained partition. The public key
of the top-level vbmeta is used as the authority in both cases as that
is immutable.

Signed-off-by: Andrew Scull <ascull@google.com>
Change-Id: I15cdb8b8997565e2a7eab5d911647a0b326e6117