ANDROID: pvmfw: Map DICE region as device memory
Instead of mapping the entire firmware region as part of the main RAM
mapping, only map the region containing DICE secrets and mark it as
device memory to avoid the need to cache-maintain it. Because its
address is only known at link time (as it depends on _end) and as
ALIGN() can't be computed at that point of the build, we instead have to
compute it at run-time and patch the corresponding mem_map entry.

Bug: 216093830
Signed-off-by: Pierre-Clément Tosi <ptosi@google.com>
Change-Id: Ib4f46f73c32157c115faf934496fbaa8ed02da3e