arm64: Do CMO_BY_VA_ONLY on every NORMAL mapping
Assume in the MMU/CMO code that the mappings provided by the board are
valid and stop filtering PTEs based on run-time RAM information during
CMOs, allowing NORMAL mappings to be properly cache-maintained, even if
found outside main RAM.

Note: As __cmo_on_leaves() and its callers are opt-in (requiring
CONFIG_CMO_BY_VA_ONLY), this change should not break existing code.
Boards wishing to enable this feature will have to ensure that their
mem_map is in a sane state before enabling the MMU/caches, which can be
performed as part of their dram_init().

Bug: 217161634
Signed-off-by: Pierre-Clément Tosi <ptosi@google.com>
Change-Id: I3a7c197582905b6c3a3487b7c5664f3b08f727e5