ANDROID: qemu-arm: Introduce CrosVM mem_map
Replace the mem_map used when building for CrosVM (determined from
SDRAM_BASE) as it is very different from the layout used by QEMU, which
may lead to undesired results. This will become particularly relevant
when using CMO_BY_VA_ONLY to flush d-cache over NORMAL regions.

Note: This is somewhat hacky. Ideally, we would want Aarch64 CrosVM to
be a U-Boot machine and the Microdroid bootloader to be a board using
that machine.

Bug: 219752211
Signed-off-by: Pierre-Clément Tosi <ptosi@google.com>
Change-Id: I9673ca78fc23493988a78462e47c0a5d24c64afc