ANDROID: qemu-arm: Introduce CrosVM mem_map
Replace the mem_map used when building for CrosVM (determined from
SDRAM_BASE) as it is very different from the layout used by QEMU, which
may lead to undesired results. This will become particularly relevant
when using CMO_BY_VA_ONLY to flush d-cache over NORMAL regions.

This re-applies 465a4ccc5d4ca2ae8708a2ef494baeaba4b64b87 but with a
fixed RAM mapping size of 255GiB.

Note: This is somewhat hacky. Ideally, we would want Aarch64 CrosVM to
be a U-Boot machine and the Microdroid bootloader to be a board using
that machine.

Bug: 219752211
Signed-off-by: Pierre-Clément Tosi <ptosi@google.com>
Change-Id: I5eec2da77a8bcb9ff8a3d162f2b249fe1bc23616