avb_verify: Simplify get_unique_guid_for_partition()
Refactor the function to use the appropriate strlcpy() (instead of a
DIY) and replace the uuid_size variable with the UUID_STR_LEN macro.
Most importantly, provide a single point of return beyond a successful
call to get_partition() to allow a future patch to apply more cleanly.

Note: No functional change intended.

Bug: 219877496
Signed-off-by: Pierre-Clément Tosi <ptosi@google.com>
Change-Id: I8b9bd748fd87592ecfcd8caf9d978cb001638277