ANDROID: pvmfw: Enroll into MMIO guard manually
Issue the MMIO guard HVC as soon as we get a chance, which has to be
before any MMIO takes place. Note that the handler for MMIO guard
injected exceptions is already in-place, as it is (indirectly) part of
vectors, initialized when VBAR is set i.e. before the C runtime is
ready. This HVC is temporarily needed until pKVM starts auto-enrolling
pVMs before boot. At that point, this commit will be reverted.

Bug: 216798684
Signed-off-by: Pierre-Clément Tosi <ptosi@google.com>
Change-Id: I1db71d79539d508178e9827859341950f5c33291