ANDROID: Indicate strict boot and new instances
Use pvmfw to indicate that the VM should strictly enforce all boot-time
checks. These strict checks are wanted in protected VMs, but may cause
problems booting non-protected VMs where the environment is slightly
different, some resources aren't available and the priority is for the
VM to boot rather than be secure.

One of these strict boot checks is to ensure that a VM instance is
provisioned during the first boot only. Checking this begins with pvmfw
which will find or create the base instance data for the VM and indicate
the status. Subsequent loading stages then enforce that provisioning
happens on the first boot of the instance and that provisioning doesn't
happen on later boots of the instance.

The device tree's chosen node is used to signal the flags between the
stages of boot.

Bug: 221051866
Bug: 217376291
Signed-off-by: Andrew Scull <ascull@google.com>
Change-Id: I5d69757eee79fbd0ae8f9ee69cb8d4e76f3ae455