firmware: psci: Add DM_FLAG_PRE_RELOC
The PSCI framework provides an API for rebooting the system and, when
available, is very likely to be the preferred (if not only) way of doing
so. In particular, it allows the PSCI driver to implement a sysreset
that U-Boot code can use during panic(). With the current DM
configuration, even if the 'psci' device tree node is available from
boot, its driver will only be instantiated after relocation (during DM
setup), resulting in calls to panic() before that point being unable to
rely on the sysreset PSCI provides.

Make the PSCI driver be probed in initf_dm() i.e. before relocation and
after only the strict minimum setup code has run in U-Boot so that code
that follows (until relocation and post-relocation before DM setup) can
properly reboot during panic(), instead of reverting to hang() when no
sysreset device seems to be available.

Bug: 222444309
Signed-off-by: Pierre-Clément Tosi <ptosi@google.com>
Change-Id: I69cfcd107688687e5646dec215bcf77922a485ec