ANDROID: Enable ns16550 driver pre-relocation
If there isn't a stdout-path in the DT's chosen node, a usable serial
devices is searched for. For this to work pre-relocation, the driver
must declare that supports running pre-relocaton.

The ns16550 does support running pre-relocation but wasn't declaring
that in order to save some memory on boards that were running out. It's
more useful to get the early logs so update the declaration as such.

Change-Id: I49d5403d3736ecb9de429f6de15a22e88d2a3d44