ANDROID: virtio: pci: Read entire capability into memory
Read the virtio PCI capability out of the device configuration space to
a struct rather than accessing fields directly from the configuration
space as they are needed. This both makes access to the fields easier
and avoids re-reading fields.

Re-reading fields could result in time-of-check to time-of-use problems,
should the value in the configuration space change. The range check of
the `bar` field and the later call to `dm_pci_read_bar32()` is an
example of where this could happen.

Bug: 217612615
Signed-off-by: Andrew Scull <ascull@google.com>
Change-Id: I7fd76e890f0b6934918a5a31fff15d5f519fc168