FROMLIST: armv8: Drop support for non-MEMORY_ATTRIBUTES MAIR
Remove the ability for caller code to set MAIR through
set_ttbr_tcr_mair() as that doesn't seem to be used nor necessary given
that the register is typically initialized once to a compile-time
constant, MEMORY_ATTRIBUTES, to be used by the hardware as a byte array
to be indexed through the 3-bit attribute entry in PTEs (see MT_NORMAL
and MT_DEVICE_*). As those attribute offsets are themselves CPP macros,
it is very unlikely for U-Boot to start dynamically defining the value
of MAIR at runtime.

For now, keep setting the register from within the function so that no
functional change should be expected.

(am from https://patchwork.ozlabs.org/project/uboot/patch/20220329170607
.689006-2-ptosi@google.com)

Signed-off-by: Pierre-Clément Tosi <ptosi@google.com>
Cc: Tom Rini <trini@konsulko.com>
Cc: Patrick Delaunay <patrick.delaunay@foss.st.com>
Cc: Patrice Chotard <patrice.chotard@foss.st.com>
Change-Id: Id648733548ac169bec732f2df6e8410b9395d55b