ANDROID: virtio_ring: Maintain a shadow copy of descriptors
The shared descriptors should only be written by the guest driver,
however, the device is still able to overwrite and corrupt them.
Maintain a private shadow copy of the descriptors for the driver to
use for state tracking, removing the need to read from the shared
descriptors.

The shadow descriptors are given the original guest address of the
buffer while the shared descriptors are given the bounce buffer address
as that is the one needed by the device.

Bug: 226232161
Signed-off-by: Andrew Scull <ascull@google.com>
Change-Id: I9a47e3fd50048b8aa70ac1170b0b678aea51fd90