scripts/dtc: Upgrade update-dtc-source.sh to v5.17
Import the script from Linux v5.17, which stops tracking the Makefile
fragments, as those aren't used by the kernel, and starts tracking
source files for YAML and DT overlay support.

Add a localized specialization of the list of files to track, to allow
U-Boot to track more files from DTC and to minimize the risk of conflict
when upgrading the script.

Use that list to keep tracking Makefile.libfdt, as it is currently used
here by pylibfdt.

Import support in the Makefile for YAML features of dtc.

Signed-off-by: Pierre-Clément Tosi <ptosi@google.com>
Change-Id: Ic226dfb80d418b08e44656b79e1a26c0e61cd650