ANDROID: Reshuffle crosvm/qemu-minimal fragments
The qemu-minimal.fragment is supposed to be a "baseline" for our virtual
platforms, as it is derived from qemu_$arch_defconfig upstream, but some
features that were not crosvm specific had started to creep into the
crosvm fragments, instead of being added to this baseline.

The justifications for BOUNCE_BUFFER and CMO_BY_VA_ONLY are perfectly
applicable to QEMU and other virtual platforms, and
SYS_MALLOC_CLEAR_ON_INIT is much needed on gem5.

Bug: 235527657
Signed-off-by: Alistair Delva <adelva@google.com>
Change-Id: I2a61c60135278885a9094a23c8cecb4e53b04b03