ANDROID: vexpress64: Fix build with CONFIG_SYSRESET=y
Some models (like gem5) emulate vexpress64 but start in EL1 by default,
and don't provide psci in EL1. This invalidates the comment in this
file, which is really obsolete anyway as CONFIG_SYSRESET has a PSCI
driver now as well. Make it possible to enable SYSRESET and bypass
this obsolete code.

Bug: 235527657
Signed-off-by: Alistair Delva <adelva@google.com>
Change-Id: Ice2c6e011c63e6e12cad57a4542cee9b6a54c05a