ANDROID: vexpress64: Support loading FDT from SYS_FDT_ADDR
The vexpress64 code did support loading an FDT, but only from flash,
and only on Juno. Update the code to also support CONFIG_SYS_FDT_ADDR
which will be checked if non-zero and Juno isn't enabled. This allows
virtual platforms based on vexpress64 to provide an FDT in a magic
RAM location, which is exactly what gem5 does when booting Linux.

Bug: 235527657
Signed-off-by: Alistair Delva <adelva@google.com>
Change-Id: I4839245c615091e7056541ab869aef76a50ae0a0