ANDROID: vexpress64: Set up fdtaddr and cbootargs
If an FDT was loaded, set up the fdtaddr environment variable with
its location in memory. If "/chosen" was specified in the FDT, also
set up the cbootargs environment variable with a copy of this string.

These features enable the fdt location and kernel cmdline to be
passed to Linux when booting the kernel.

This feature is similar to the QEMU implementation, but so as to not
break vexpress64 FVT with no FDT, check that the FDT could be loaded
before proceeding with the setup.

Bug: 235527657
Signed-off-by: Alistair Delva <adelva@google.com>
Change-Id: Ida77dfbdc208935c6a32ad7ba2c70e56d469d7af