pvmfw: Map FDT in kernel boot mode
pvmfw currently assumes that the FDT is at the base of RAM, matching the
BIOS boot mode of crosvm. This doesn't hold for kernel boot mode, where
it is placed at RAM_TOP-2MB-4K.

To this end, create a new mmap entry for the FDT, populated only if the
FDT pointer is not at RAM base. Because of CONFIG_CMO_BY_VA_ONLY=y, this
new region is not invalidated/flushed together with the rest of RAM.
Cache maintenance is therefore done manually.

FDT placed inside the 32M SDRAM region is not supported.

Bug: 244294566
Test: boots in both bios and kernel modes
Change-Id: I7f212fbfbcb70e801c6f8891835013af2591004e