ANDROID: Fix alignment of DTB blob in u-boot-dtb.bin
Compared to upstream, we use a newer version of libfdt from the dtc
project via a symlink, but this newer code requires the dtb to be
aligned to 8 bytes, not 4 bytes, and upstream's linker script can't
handle this.

So, move the DTB to be aligned to 8 bytes, by updating the alignment
of the _end address set by the linker script, and extending the size
of the nodtb.bin on x86 only to be a multiple of 8 bytes. I couldn't
find a cleaner way to do it, as objcopy -O binary only copies sections
and not symbols, and the size of the last section copied matters and
can't just be hacked to be a multiple of 8 bytes..

This was randomly breaking either crosvm or QEMU prebuilts for x86
because they would rarely be aligned the same.

Bug: 250067610
Signed-off-by: Alistair Delva <adelva@google.com>
Change-Id: I68f1dd3dbd14c08d53859c8eebfafea5440ada6e