start.S: Fix compilation issue
Approach with Bazel and repo, which is used by Google to compile U-Boot
is using a clang compiler instead of gcc. That compiler treats
redefinition of the '_start' symbol as error.

This patch is intended to fix a clang-specific compilation error. The
'asm/boot0-linux-kernel-header.h' header defines '_start" symbol by
itself. Thus, code should be slightly changed to avoid symbol
redefinition in case of 'CONFIG_LINUX_KERNEL_IMAGE_HEADER' flag is
defined.

Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>
Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>